Solution to Calculate Average Temperature of a Week

+++ Calculate Average Temperature of a Week@@67877d61dc107486ad5c274d/code.c
@@ -2,12 +2,12 @@
 
 struct week{
     int days;
-    float tempp;
+    int tempp;
 };
 
 int main() {
     struct week e[7];
-    int temp=0;
+    float temp=0;
     float average;
     for(int i=0;i<7;i++){
         scanf("%d",e[i].days);
